+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Pattern/PatternBase.cs
@@ -6,35 +6,35 @@ using System.Threading;
 public abstract class PatternBase : ScriptableObject
 {
     [Tooltip("ExecuteImplの実行前に待機する時間（秒）")]
-    [SerializeField] private float _beforeAwaitSeconds = 0f;
+    [SerializeField] private FloatReference _beforeAwaitSeconds = new FloatReference { useConstant = true, constantValue = 0f };
 
     [Tooltip("ExecuteImplの実行後に待機する時間（秒）")]
-    [SerializeField] private float _afterAwaitSeconds = 0f;
+    [SerializeField] private FloatReference _afterAwaitSeconds = new FloatReference { useConstant = true, constantValue = 0f };
 
     // GameEntityControllerを受け取るExecuteメソッド
     public virtual async UniTask Execute(GameEntityController controller, CancellationToken token)
     {
         if (token.IsCancellationRequested) return;
-        if (_beforeAwaitSeconds > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_beforeAwaitSeconds), cancellationToken: token);
+        if (_beforeAwaitSeconds.Value > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_beforeAwaitSeconds.Value), cancellationToken: token);
         if (token.IsCancellationRequested) return;
 
         await ExecuteImpl(controller, token);
 
         if (token.IsCancellationRequested) return;
-        if (_afterAwaitSeconds > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_afterAwaitSeconds), cancellationToken: token);
+        if (_afterAwaitSeconds.Value > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_afterAwaitSeconds.Value), cancellationToken: token);
     }
 
     // MovementStateを受け取るExecuteメソッドのオーバーロード
     public virtual async UniTask Execute(MovementState state, CancellationToken token)
     {
         if (token.IsCancellationRequested) return;
-        if (_beforeAwaitSeconds > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_beforeAwaitSeconds), cancellationToken: token);
+        if (_beforeAwaitSeconds.Value > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_beforeAwaitSeconds.Value), cancellationToken: token);
         if (token.IsCancellationRequested) return;
 
         await ExecuteImpl(state, token);
 
         if (token.IsCancellationRequested) return;
-        if (_afterAwaitSeconds > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_afterAwaitSeconds), cancellationToken: token);
+        if (_afterAwaitSeconds.Value > 0) await UniTask.Delay(System.TimeSpan.FromSeconds(_afterAwaitSeconds.Value), cancellationToken: token);
     }
 
     // サブクラスで具体的な処理を実装するための抽象メソッド